+++ yt/ytlib/chunk_client/file_chunk_writer.h
@@ -61,7 +61,7 @@ public:
         // Check alignment.
         YASSERT(
             static_cast<i64>(reinterpret_cast<char*>(infoEnd) - reinterpret_cast<char*>(infoBegin)) ==
-            footer.BlockCount * sizeof (TBlockInfo));
+            static_cast<i64>(footer.BlockCount * sizeof (TBlockInfo)));
 
         File->Write(infoBegin, footer.BlockCount * sizeof (TBlockInfo));
